# openapi.yaml
openapi: 3.1.0
info:
  title: Codex Artifacts publishing API
  version: 0.1.0
  description: Owner automation API for one self-hosted Codex Artifacts deployment.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://codex-artifacts.your-account.workers.dev
paths:
  /api/status:
    get:
      operationId: getStatus
      summary: Check service liveness
      responses:
        "200":
          description: Service process is responding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
  /api/artifacts:
    post:
      operationId: publishArtifact
      summary: Create an artifact or replace the artifact with a requested slug
      description: |
        The bearer token is deployment-wide owner automation authority. If `slug`
        identifies an existing artifact, its HTML and title are replaced. Omitted
        `sharedWith`, `sharedDomains`, and `isPublic` values preserve existing
        settings during updates.
        Omitted `expiresInSeconds` values start or reset the three-day expiration
        timer. Set it to `null` for an artifact that never expires.
        Deployment-configured workspace viewers retain read access independently
        of `sharedWith`.
      security:
        - publishToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "200":
          description: Existing artifact updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
        "201":
          description: Artifact created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
        "400":
          description: Invalid artifact input or exhausted workspace capacity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid publishing token.
          content:
            text/plain:
              schema:
                type: string
        "503":
          description: Publishing automation is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    publishToken:
      type: http
      scheme: bearer
  schemas:
    Status:
      type: object
      additionalProperties: false
      required: [ok, service]
      properties:
        ok:
          type: boolean
          const: true
        service:
          type: string
          const: codex-artifacts
    PublishRequest:
      type: object
      additionalProperties: false
      required: [title, html]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        slug:
          type: string
          description: Optional stable slug. Non-alphanumeric runs normalize to hyphens.
          maxLength: 80
        html:
          type: string
          description: Self-contained UTF-8 HTML, limited to 512 KiB by the server.
        sharedWith:
          type: array
          maxItems: 50
          description: |
            Additional artifact-specific exact-email recipients. Replaces only
            this list when updating; workspace viewers retain access.
          items:
            type: string
            format: email
        sharedDomains:
          type: array
          maxItems: 20
          description: |
            Additional artifact-specific email domains. Replaces only this
            list when updating; workspace viewers retain access.
          items:
            type: string
            format: hostname
        isPublic:
          type: boolean
        expiresInSeconds:
          description: |
            Relative lifetime in whole seconds, from 60 seconds through one year.
            Omit for the three-day default or set to null to never expire. The
            lifetime starts again whenever the artifact is published or replaced.
          oneOf:
            - type: integer
              minimum: 60
              maximum: 31536000
            - type: "null"
    PublishResponse:
      type: object
      additionalProperties: false
      required: [id, slug, updated, isPublic, expiresAt]
      properties:
        id:
          type: string
        slug:
          type: string
        updated:
          type: boolean
        isPublic:
          type: boolean
        expiresAt:
          description: Absolute expiration time, or null when expiration is disabled.
          oneOf:
            - type: string
              format: date-time
            - type: "null"
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string